Do not ever fetch data when refreshing the state of the sensors
Sometimes, it will fire an API request when attempting to refresh
the state of the sensors. This routine is done every minute in order
to make sure that the power sensors always have the "earliest"
data but fails.

This is most likely why some users are facing rate-limits, coupled
with having neighbors that also use Open-Meteo. It explains a lot.

Signed-off-by: rany <ranygh@riseup.net>